+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v4.8.0
 * Fixed: Soft delete with event data cannot be un-soft deleted
 * Changed: Elasticsearch to use bulk updates
+* Changed: EdgeVertexPairs can now have null vertex ids if user can see the edge but not the vertex
 
 # v4.7.1
 * Added: Compare.RANGE to query